+++ src/main/java/camp/nextstep/edu/kitchenpos/order/domain/OrderLineItem.java
@@ -1,4 +1,4 @@
-package camp.nextstep.edu.kitchenpos.order.model;
+package camp.nextstep.edu.kitchenpos.order.domain;
 
 public class OrderLineItem {
     private Long seq;
